+++ mern/chompquest-frontend/src/components/ProfilePicture.css
@@ -1,0 +1,64 @@
+.profile-picture-container {
+  position: fixed; 
+  top: 12px;
+  right: 16px;
+  z-index: 2000; 
+  background: none;
+  display: flex;
+  align-items: center;
+}
+
+.profile-icon-btn {
+  border: none;
+  padding: 0;
+  cursor: pointer;
+  display: flex; 
+  align-items: center;
+  justify-content: center;
+  width: 60px; 
+  height: 60px; 
+  border-radius: 50%;
+  box-shadow: 0 2px 5px rgba(0,0,0,0.4);
+  transition: transform 0.1s ease;
+}
+
+.profile-icon-btn:hover {
+  transform: scale(1.05);
+}
+
+.profile-dropdown-menu {
+  position: absolute; 
+  top: 60px;
+  right: 0;
+  background: #fff;
+  border: 1px solid #ddd;
+  border-radius: 8px;
+  box-shadow: 0 4px 16px rgba(0,0,0,0.13);
+  min-width: 120px;
+  font-size: 13px;
+  padding: 4px 0;
+  z-index: 2001; 
+}
+
+.profile-dropdown-item {
+  width: 100%;
+  padding: 7px 14px;
+  background: none;
+  border: none;
+  text-align: left;
+  cursor: pointer;
+  font-size: 13px;
+  transition: background-color 0.15s ease;
+}
+
+.profile-dropdown-item:hover {
+  background-color: #f5f5f5;
+}
+
+.profile-dropdown-item.logout {
+  color: #d32f2f;
+}
+
+.profile-dropdown-item.logout:hover {
+  background-color: #ffebee;
+}
