+++ mern/chompquest-frontend/src/components/ProgressCard.css
@@ -1,0 +1,55 @@
+.progress-card-container {
+  position: fixed;
+  bottom: 32px;
+  right: 32px;
+  z-index: 1000;
+}
+
+.progress-card {
+  width: 290px;
+  height: 275px;
+  background: #fff;
+  border-radius: 16px;
+  box-shadow: 0 2px 16px rgba(0,0,0,0.08);
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  font-family: inherit;
+}
+
+.progress-card-title {
+  font-size: 1.35rem;
+  font-weight: 600;
+  color: #222;
+  margin-bottom: 18px;
+  width: 100%;
+  text-align: center;
+}
+
+.progress-card-medal {
+  display: flex;
+  justify-content: center;
+  margin-bottom: 18px;
+}
+
+.progress-card-medal img {
+  width: 100px;
+  height: 100px;
+  border-radius: 50%;
+}
+
+.progress-card-rank {
+  font-weight: 700;
+  color: #b08d57;
+  margin-bottom: 10px;
+  font-size: 1.15rem;
+  text-align: center;
+}
+
+.progress-card-stats {
+  color: #444;
+  opacity: 0.95;
+  font-size: 1.05rem;
+  text-align: center;
+}
